esp32-cloud-backend/requirements.txt
flask
supabase
bcrypt
PyJWT